worker/requirements.txt
openai
redis
azure-storage-blob
azure-servicebus
sqlalchemy
asyncpg
azure-identity
PyPDF2
python-docx